Post-processing APSIM outputs into netcdf files

Post-processing APSIM outputs into netcdf files

apsimOut2Nc(
  filename,
  years,
  factors,
  traits,
  par_factor,
  base,
  factor_files,
  base_sub = NULL,
  merge_out_nc = TRUE,
  yearvar = "year",
  cpu_num = 14,
  tmp_folder = file.path("Results", "tmpNc")
)

Arguments

filename

The filename of netcdf file

years

The years in the APSIM and netcdf

factors

The factors in the netcdf and APSIM

traits

The vector of traits write into netcdf file

par_factor

The factors will be parallel processed

base

The base folder to store APSIM output files

factor_files

The data.frame or vector of files to specify the output file names and levels

base_sub

The sub-folder to store the single output files

merge_out_nc

Whether merge the parallel outputs into an array when updating netcdf files.

yearvar

Variable name of yeas in the APSIM output file

cpu_num

The cpu number of parallel

tmp_folder

The tmp folder to store merge APSIM output files