Preparing task table for APSIM simulations

Preparing task table for APSIM simulations

clusterRunPreparingTask(
  project,
  factors,
  sharedrive,
  local,
  par_factor,
  input_factor,
  skip_factor = NULL,
  folder_prefix = NULL,
  r_script = "R\\bin\\i386\\Rscript.exe",
  r_code = "RCode.R",
  ids = 0,
  save_factor = FALSE
)

Arguments

project

The project name in the ClusterRun

factors

The factors list to generate simulations

sharedrive

sharedrive

local

local

par_factor

The parallel factors to write into database of ClusterRun

input_factor

The factors to indicate input files

skip_factor

The skip factors to be processed in the condor clients

folder_prefix

Output folder to store the out files of all simulations

r_script

The path to Rscript.exe file

r_code

r_code

ids

The starting id of simulations